Time: 14 ms (22.34%), Space: 8.6 MB (33.69%) - LeetHub

+++ 0009-palindrome-number/0009-palindrome-number.cpp
@@ -1,8 +1,8 @@
 class Solution {
 public:
-    int recursive(int x, int res){
+    int recursive(int x, long long res){
         if(x==0) return res;
-        res = res*10 + x%10;
+        res = res*10LL + x%10;
         return recursive(x/10, res);
     }
     bool isPalindrome(int x) {
